move `Cell` conversions to new module & update
